Import all Wendigo Voice lines, Add WendigoLevel dialogue, Add dialogue in script
Something is triggering the _on_forest_bounds_body_entered signals whenever the game starts causing both the flashlight and beginning dialogues to overlap.
also, there is a weird glitchy double screen at the bottom. Thought it was the old dialogue in the UI, but I deleted that and it still remains. Might be related to the gray bar at the bottom of the title screen?